requirements.txt
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
click==8.1.7
cryptography==44.0.0
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.6
filelock==3.16.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.3
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
nodeenv==1.9.1
passlib==1.7.4
platformdirs==4.3.6
pre_commit==4.0.1
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.3
typer==0.15.1
types-passlib==1.7.7.20240819
typing_extensions==4.12.2
uvicorn==0.33.0
uvloop==0.21.0
virtualenv==20.28.0
watchfiles==1.0.3
websockets==14.1
